refactor: move FlowExt to network module and add cancellable call
The `FlowExt.kt` file has been moved from the `ui` module to the `network` module. This change centralizes network-related extensions within the `network` module. Additionally, a `cancellable()` call has been added within the catch block to ensure that cancellation exceptions are not suppressed. Based on the file extension and content, both files seems to be a .kt file that is normal source file not script file.